+++ GE-Tre-PIP/GE-Tre-PIP/GQT.h
@@ -50,7 +50,7 @@ public:
 	double caculateDistance(Point p, Node* n);
 	double caculateDistance_stat(Point p, Node* n);
 	Statistic& getStat();
-	void GEtree_statStorageCost(int length, int widthint, int* auxiliary_cost);
+	void GEtree_statStorageCost(int length, int widthint, long long int* auxiliary_cost);
 };
